fix: tighten shortcuts lookup and split modules
