+++ crates/adapter-macos/src/shortcuts/compiler/mod.rs
@@ -1,0 +1,5 @@
+mod decode;
+mod encode;
+
+pub use decode::decompile_actions;
+pub use encode::{append_action, compile_actions};
