+++ crates/cli/Cargo.toml
@@ -18,5 +18,6 @@ cueward-core = { workspace = true }
 cueward-adapter-macos = { workspace = true }
 clap = { version = "4", features = ["derive"] }
 chrono = "0.4"
+serde = "1"
 serde_json = "1"
 serde_yaml = "0.9"
