+++ crates/cli/src/commands/shortcuts/mod.rs
@@ -1,0 +1,6 @@
+mod args;
+mod dispatch;
+mod io;
+
+pub(crate) use args::ShortcutsAction;
+pub(crate) use dispatch::dispatch;
